requirements.txt
# ================================
# WEB FRAMEWORK Y SERVIDOR
# ================================

# FastAPI - Framework web moderno para crear APIs REST
# Características: validación automática, documentación auto-generada, type hints
# Actualizado a 0.125.0 para resolver CVE-2024-47874 y CVE-2025-54121 en starlette
fastapi==0.125.0

# Uvicorn - Servidor ASGI de alto rendimiento para ejecutar aplicaciones FastAPI
# [standard] incluye extras: watchfiles (auto-reload), websockets, colorama (logs coloridos)
uvicorn[standard]==0.30.0


# ================================
# CONFIGURACIÓN Y ENTORNO
# ================================

# python-dotenv - Utilidad para cargar variables de entorno desde un fichero .env
# Útil para gestionar la configuración en desarrollo local sin Docker
python-dotenv==1.0.1


# ================================
# DATABASE & ORM
# ================================

# SQLAlchemy - El ORM (Object-Relational Mapper) más potente de Python
# Permite interactuar con la base de datos usando objetos Python en lugar de SQL
sqlalchemy==2.0.31

# psycopg2-binary - Adaptador (driver) para conectar SQLAlchemy con PostgreSQL
# Es la librería puente entre nuestra aplicación y la base de datos
psycopg2-binary==2.9.9

# Alembic - Herramienta para migraciones de bases de datos con SQLAlchemy
# Permite versionar y aplicar cambios en el esquema de la base de datos
alembic==1.13.2

# asyncpg - Driver asíncrono para PostgreSQL, necesario para SQLAlchemy en modo async
# Es la alternativa a psycopg2 para operaciones no bloqueantes con la base de datos
asyncpg==0.29.0

# greenlet - Librería para concurrencia ligera (micro-hilos)
# Requerido por SQLAlchemy para puentear sus componentes internos síncronos en un entorno async
greenlet==3.0.3


# ================================
# SEGURIDAD
# ================================

# bcrypt - Librería para hashing de contraseñas
# Implementa el algoritmo bcrypt, un estándar para el almacenamiento seguro de claves
bcrypt==4.2.0

# python-jose - Librería para crear y verificar tokens JWT (JSON Web Tokens)
# Incluye soporte para varios algoritmos de firma (HS256, RS256, etc.)
# Actualizado a 3.4.0+ para resolver vulnerabilidad de seguridad CVE-2024-XXXXX
python-jose[cryptography]==3.4.0

# slowapi - Rate limiting para FastAPI
# Protege contra brute force, DoS y abuso de API limitando peticiones por IP
# Basado en Flask-Limiter, soporta memoria local y Redis como backend
slowapi==0.1.9

# secure - Security Headers HTTP para FastAPI
# Añade automáticamente headers de seguridad (CSP, HSTS, X-Frame-Options, etc.)
# Protege contra XSS, Clickjacking, MIME-sniffing, MITM attacks
# Cumple con OWASP Top 10: A02, A03, A04, A05, A07
secure==0.3.0

# sentry-sdk - Error tracking y performance monitoring
# Captura automáticamente excepciones, errores y métricas de rendimiento
# Integraciones: FastAPI, SQLAlchemy, Redis
# Cumple con OWASP Top 10: A09 (Logging & Monitoring)
sentry-sdk[fastapi]==2.19.2

# safety - Auditoría de vulnerabilidades en dependencias
# Escanea el proyecto contra Safety DB (40K+ vulnerabilidades conocidas)
# Detecta CVEs en dependencias directas y transitivas
# Cumple con OWASP Top 10: A06 (Vulnerable and Outdated Components)
safety==3.7.0

# pip-audit - Auditoría oficial de PyPA para dependencias
# Escanea contra PyPI Advisory Database y OSV (Open Source Vulnerabilities)
# Alternativa oficial mantenida por Python Packaging Authority
# Cumple con OWASP Top 10: A06 (Vulnerable and Outdated Components)
pip-audit==2.10.0

# urllib3 - Cliente HTTP de bajo nivel (dependencia transitiva de requests/httpx)
# Fijada explícitamente para resolver:
# - CVE-2025-66418 y CVE-2025-66471 (decompression bomb y unlimited decompression chain)
# - CVE-2026-21441 (Security vulnerability - fixed in 2.6.3)
urllib3==2.6.3

# filelock - Sincronización de archivos mediante locks (dependencia transitiva de virtualenv/safety)
# Fijada explícitamente para resolver CVE-2025-68146 (TOCTOU race condition)
filelock==3.20.1

# authlib - OAuth and OpenID Connect library (dependencia transitiva de safety)
# Fijada explícitamente para resolver:
# - CVE-2025-61920 (DoS via large base64 tokens)
# - CVE-2025-62706 (DoS via zip bomb decompression)
# - CVE-2025-68158 (Login CSRF vulnerability - fixed in 1.6.6)
authlib==1.6.6

# setuptools - Build system y package installer (dependencia transitiva de safety)
# Fijada explícitamente para resolver:
# - CVE-2024-6345 (Code Injection via package_index - CVSS 7.5 HIGH)
# - CVE-2022-40897 (Regular Expression DoS - CVSS 5.9 MEDIUM)
# - CVE-2025-47273 (Directory Traversal en _download_url - CVSS 6.8 MEDIUM)
setuptools==78.1.1

# zipp - Pathlib-compatible Zipfile object wrapper (dependencia transitiva de importlib-metadata)
# Fijada explícitamente para resolver CVE-2024-5569 (Infinite loop DoS via Path module)
zipp==3.19.1

# marshmallow - Object serialization/deserialization library (dependencia transitiva de safety)
# Fijada explícitamente para resolver CVE-2025-68480 (DoS via Asymmetric Resource Consumption)
marshmallow==3.26.2

# ecdsa - Elliptic Curve Digital Signature Algorithm (dependencia transitiva de python-jose)
# Fijada explícitamente para resolver CVE-2024-23342 (Signature malleability vulnerability)
# NOTA: No hay versión fija disponible aún, monitorear actualizaciones
ecdsa==0.19.1


# ================================
# TESTING Y DESARROLLO
# ================================

# email-validator - Librería robusta para validar direcciones de email
# Usada en los Value Objects para asegurar la validez de los emails
email-validator==2.3.0

# httpx - Cliente HTTP asíncrono para testear endpoints de FastAPI
# Alternativa moderna a `requests` con soporte para async/await
httpx==0.27.0

# requests - Cliente HTTP síncrono estándar de Python
# Utilizado para integración con servicios externos (Mailgun, RFEG)
requests==2.32.4

# pytest - Framework de testing moderno y potente
# Más simple que unittest, mejor output de errores, fixtures potentes
pytest==8.3.0

# pytest-json-report - Plugin para pytest que genera reportes en formato JSON
# Esencial para el análisis programático de los resultados de los tests
pytest-json-report==1.5.0

# pytest-xdist - Plugin para pytest que permite ejecutar tests en paralelo
# Acelera significativamente la suite de tests distribuyendo el trabajo
pytest-xdist==3.8.0

# pytest-asyncio - Plugin para pytest que proporciona soporte para tests asíncronos
# Permite usar async/await en tests y fixtures con asyncio_mode = auto
pytest-asyncio==0.24.0
pytest-cov
